splitting mathics.builtin.patterns in smaller submodules
